Organize (move) client side CJ related code
