+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/BobClient.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using WabiSabi.Crypto.ZeroKnowledge;
 
-namespace WalletWasabi.WabiSabi.Client;
+namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client;
 
 public class BobClient
 {
