+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinClient.cs
@@ -22,7 +22,7 @@ using WalletWasabi.WabiSabi.Models;
 using WalletWasabi.WabiSabi.Models.MultipartyTransaction;
 using WalletWasabi.WebClients.Wasabi;
 
-namespace WalletWasabi.WabiSabi.Client;
+namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client;
 
 public class CoinJoinClient
 {
@@ -617,8 +617,7 @@ public class CoinJoinClient
 		var maximumSigningRequestDelay = TimeSpan.FromSeconds(50);
 		var scheduledDates = signingEndTime.GetScheduledDates(aliceClients.Count(), signingStartTime, maximumSigningRequestDelay);
 
-		var tasks = Enumerable.Zip(
-			aliceClients,
+		var tasks = aliceClients.Zip(
 			scheduledDates,
 			async (aliceClient, scheduledDate) =>
 			{
@@ -648,8 +647,7 @@ public class CoinJoinClient
 	{
 		var scheduledDates = readyToSignEndTime.GetScheduledDates(aliceClients.Count(), DateTimeOffset.UtcNow, MaximumRequestDelay);
 
-		var tasks = Enumerable.Zip(
-			aliceClients,
+		var tasks = aliceClients.Zip(
 			scheduledDates,
 			async (aliceClient, scheduledDate) =>
 			{
@@ -741,7 +739,7 @@ public class CoinJoinClient
 
 		// Splitting the remaining time.
 		// Both operations are done under output registration phase, so we have to do the random timing taking that into account.
-		var outputRegistrationEndTime = now + (remainingTime * 0.8); // 80% of the time.
+		var outputRegistrationEndTime = now + remainingTime * 0.8; // 80% of the time.
 		var readyToSignEndTime = outputRegistrationEndTime + remainingTime * 0.2; // 20% of the time.
 
 		CoinJoinClientProgress.SafeInvoke(this, new EnteringOutputRegistrationPhase(roundState, outputRegistrationPhaseEndTime));
@@ -842,7 +840,7 @@ public class CoinJoinClient
 	private async Task<ImmutableArray<(AliceClient, PersonCircuit)>> ProceedWithInputRegAndConfirmAsync(IEnumerable<SmartCoin> smartCoins, RoundState roundState, CancellationToken cancellationToken)
 	{
 		// Because of the nature of the protocol, the input registration and the connection confirmation phases are done subsequently thus they have a merged timeout.
-		var timeUntilOutputReg = (roundState.InputRegistrationEnd - DateTimeOffset.UtcNow) + roundState.CoinjoinState.Parameters.ConnectionConfirmationTimeout;
+		var timeUntilOutputReg = roundState.InputRegistrationEnd - DateTimeOffset.UtcNow + roundState.CoinjoinState.Parameters.ConnectionConfirmationTimeout;
 
 		using CancellationTokenSource timeUntilOutputRegCts = new(timeUntilOutputReg + ExtraPhaseTimeoutMargin);
 		using CancellationTokenSource linkedCts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, timeUntilOutputRegCts.Token);
