+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinProgressEvents/CoinJoinProgressEventArgs.cs
