+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinProgressEvents/EnteringConnectionConfirmationPhase.cs
