+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinProgressEvents/EnteringCriticalPhase.cs
