+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinProgressEvents/EnteringSigningPhase.cs
