+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinProgressEvents/LeavingCriticalPhase.cs
