+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/CoinJoinResult.cs
