+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/Decomp/DenominationBuilder.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using WabiSabi.Crypto.Randomness;
 using WalletWasabi.Extensions;
 
-namespace WalletWasabi.WabiSabi.Client;
+namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomp;
 
 public static class DenominationBuilder
 {
