+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/Decomp/Output.cs
@@ -2,7 +2,7 @@ using NBitcoin;
 using System.Diagnostics.CodeAnalysis;
 using WalletWasabi.Extensions;
 
-namespace WalletWasabi.WabiSabi.Client;
+namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomp;
 
 public record Output
 {
