+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/CoinRefrigerator.cs
