+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/CoinJoinStatusEventArgs.cs
