+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/CompletedEventArgs.cs
