+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/LoadedEventArgs.cs
