+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/StartErrorEventArgs.cs
