+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/StartedEventArgs.cs
