+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/StatusChangedEventArgs.cs
