+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/StoppedEventArgs.cs
