+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/WalletStartedCoinJoinEventArgs.cs
