+++ WalletWasabi/WabiSabi/Client/CoinJoin/Manager/StatusChangedEvents/WalletStoppedCoinJoinEventArgs.cs
