+++ WalletWasabi/WabiSabi/Client/SmartRequestNode.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using WabiSabi.Crypto.ZeroKnowledge;
+using WalletWasabi.WabiSabi.Client.CoinJoin.Client;
 
 namespace WalletWasabi.WabiSabi.Client;
